app/requirements.txt
fastapi
uvicorn
langchain==0.3.8
sentence-transformers
huggingface-hub
requests
together
pyngrok
nest-asyncio
langchain-community
langchain-chroma
python-dotenv
boto3
langchain-huggingface
tf-keras
